feat: new UI rendering component for ESC
